backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-chroma
chromadb
langchain-huggingface
pydantic
python-multipart
